requirement.txt
SpeechRecognition
ivy-python
PyAudio
opencv-python
mediapipe
numpy